+++ v3/internal/templates/sveltekit-ts/frontend/vite.config.ts
@@ -1,0 +1,16 @@
+import { sveltekit } from '@sveltejs/kit/vite';
+import { defineConfig, searchForWorkspaceRoot } from 'vite';
+
+export default defineConfig({
+    server: {
+        fs: {
+          allow: [
+            // search up for workspace root
+            searchForWorkspaceRoot(process.cwd()),
+            // your custom rules
+            './bindings/*',
+          ],
+        },
+    },
+	plugins: [sveltekit()]
+});