+++ v3/internal/templates/sveltekit/frontend/src/routes/+layout.js
@@ -1,0 +1,2 @@
+export const prerender = true;
+export const ssr = false;